started ast
