+++ fastn-js/tests/03-event-1.html
@@ -12,7 +12,7 @@
             let num = fastn.mutable(4);
             let value = fastn.formula([num], function () { return num.get() * 10; })
             let i = fastn_dom.createKernel(root, fastn_dom.ElementKind.Integer);
-            i.setDynamicProperty(fastn_dom.PropertyKind.IntegerValue, [value], function () { return value.get(); });
+            i.setProperty(fastn_dom.PropertyKind.IntegerValue, value);
             i.addEventHandler(fastn_dom.Event.Click, function () {
                 num.set(num.get() + 1);
             });
